+++ pyogrio/util.py
@@ -4,11 +4,14 @@ import re
 import sys
 from packaging.version import Version
 from pathlib import Path
+from typing import Union
 from urllib.parse import urlparse
+
+from pyogrio._vsi import vsimem_rmtree_toplevel as _vsimem_rmtree_toplevel
 
 
 def get_vsi_path_or_buffer(path_or_buffer):
-    """Get vsi-prefixed path or bytes buffer depending on type of path_or_buffer.
+    """Get VSI-prefixed path or bytes buffer depending on type of path_or_buffer.
 
     If path_or_buffer is a bytes object, it will be returned directly and will
     be read into an in-memory dataset when passed to one of the Cython functions.
@@ -29,9 +32,10 @@ def get_vsi_path_or_buffer(path_or_buffer):
     str or bytes
 
     """
-    # force path objects to string to specifically ignore their read method
+    # treat Path objects here already to ignore their read method + to avoid backslashes
+    # on Windows.
     if isinstance(path_or_buffer, Path):
-        return vsi_path(str(path_or_buffer))
+        return vsi_path(path_or_buffer)
 
     if isinstance(path_or_buffer, bytes):
         return path_or_buffer
@@ -48,13 +52,14 @@ def get_vsi_path_or_buffer(path_or_buffer):
     return vsi_path(str(path_or_buffer))
 
 
-def vsi_path(path: str) -> str:
-    """Ensure path is a local path or a GDAL-compatible vsi path."""
-    if "/vsimem/" in path:
-        raise ValueError(
-            "path cannot contain /vsimem/ directly; to use an in-memory dataset a "
-            "bytes object must be passed instead"
-        )
+def vsi_path(path: Union[str, Path]) -> str:
+    """Ensure path is a local path or a GDAL-compatible VSI path."""
+    # Convert Path objects to string, but for VSI paths, keep posix style path.
+    if isinstance(path, Path):
+        if sys.platform == "win32" and path.as_posix().startswith("/vsi"):
+            path = path.as_posix()
+        else:
+            path = str(path)
 
     # path is already in GDAL format
     if path.startswith("/vsi"):
@@ -217,3 +222,26 @@ def _mask_to_wkb(mask):
         raise ValueError("'mask' parameter must be a Shapely geometry")
 
     return shapely.to_wkb(mask)
+
+
+def vsimem_rmtree_toplevel(path: Union[str, Path]):
+    """Remove the parent directory of the file path recursively.
+
+    This is used for final cleanup of an in-memory dataset, which may have been
+    created within a directory to contain sibling files.
+
+    Additional VSI handlers may be chained to the left of /vsimem/ in path and
+    will be ignored.
+
+    Remark: function is defined here to be able to run tests on it.
+
+    Parameters
+    ----------
+    path : str or pathlib.Path
+        path to in-memory file
+
+    """
+    if isinstance(path, Path):
+        path = path.as_posix()
+
+    _vsimem_rmtree_toplevel(path)