+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## 0.10.0 (yyyy-mm-dd)
 
+### Improvements
+
+-   Add support to read, write, list, and remove `/vsimem/` files (#457)
+
 ### Bug fixes
 
 -   Silence warning from `write_dataframe` with `GeoSeries.notna()` (#435).
@@ -18,6 +22,8 @@
     This also leads to `pyproj` becoming an optional dependency; you will need
     to install `pyproj` in order to support spatial reference systems (#452).
 -   The GDAL library included in the wheels is updated from 3.8.5 to GDAL 3.9.2 (#466).
+-   pyogrio now requires a minimum version of Python >= 3.9 (#473).
+-   Wheels are now available for Python 3.13.
 
 ## 0.9.0 (2024-06-17)